refactor: move useReferenceAutoCompletePopoverBoundary hook to proper dir
